third_party/WebKit/LayoutTests/external/wpt/html-media-capture/capture_reflect-expected.txt

This is a testharness.js-based test.
FAIL Element input should have own property captureassert_true: expected true got false
FAIL input.capture is "" when the capture attribute is absentassert_equals: expected (string) "" but got (undefined) undefined
FAIL input.capture is "" when the capture attribute is missing value defaultassert_equals: expected (string) "" but got (undefined) undefined
FAIL input.capture is "user" when the capture attribute is userassert_equals: expected (string) "user" but got (undefined) undefined
FAIL input.capture is "invalid" when the capture attribute is invalid value defaultassert_equals: expected (string) "invalid" but got (undefined) undefined
FAIL input.capture is "environment" when the capture attribute is environmentassert_equals: expected (string) "environment" but got (undefined) undefined
Harness: the test ran to completion.
